refactor(types): элементы и таблицы формы вынесены в свой регистратор
Второй разрез `FormTypesProvider` (`java:S1448`, `java:S1200`):
`FormItemTypesRegistrar` забрал типы видов элементов, коллекцию `Элементы`, типы
таблиц и правила их членов. Провайдер — 1524 → 1050 строк.

Общее для регистраторов вынесено по смыслу: заведение синтетического типа с
английским алиасом — в маленький `FormTypeFactory`, свойство платформенного вида
(`platformProperty`) — в словарь имён `FormPlatformTypes`.
